+++ Dogs.Backend/src/Core/Dogs.Core/ValueObjects/TailLenght.cs
@@ -5,17 +5,16 @@ namespace Dogs.Core.ValueObjects
 {
     public class TailLenght
     {
-        public int Value { get; }
-        public TailLenght(int value)
+        public double Value { get; }
+        public TailLenght(double value)
         {
             if (value is < 0 or > 150)
                 throw new InvalidTailLenghtException(value);
 
             Value = value;
         }
-        public static implicit operator TailLenght(int value) => new(value);
+        public static implicit operator TailLenght(double value) => new(value);
 
-        public static implicit operator int(TailLenght value) => value.Value;
-
+        public static implicit operator double(TailLenght value) => value.Value;
     }
 }
